[mod]gcc编译器下的代码兼容性修改

+++ src/WtDataReader/WtDataReader.cpp
@@ -2152,7 +2152,7 @@ WTSKlineSlice* WtDataReader::readKlineSlice(const char* stdCode, WTSKlinePeriod 
 		hisCnt = left;
 		//历史数据，直接从缓存的历史数据尾部截取
 		BarsList& barList = _bars_cache[key];
-		hisCnt = min(hisCnt, barList._bars.size());
+		hisCnt = min(hisCnt, (uint32_t)barList._bars.size());
 		hisHead = &barList._bars[barList._bars.size() - hisCnt];//indexBarFromCache(key, etime, hisCnt, period == KP_DAY);
 	}
